postgresql/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.5.1
python-multipart>=0.0.7
psycopg2-binary==2.9.9
